+++ sdk/src/components/action-barcode.tsx
@@ -1,10 +1,10 @@
+import { ComponentChildren } from "preact";
 import { useCallback, useMemo, useState } from "preact/hooks";
 import { amountFormat } from "../amount-format";
-import { Instructions as InstructionsType } from "../backend-types/instructions";
 import { TFunction } from "../localization";
 import { generateBarcodeSvg } from "./action-barcode-utils";
+import { useActionCard } from "./action-card";
 import { Button, ButtonLoadingSpinner, ButtonVariant } from "./core/button";
-import { Instructions } from "./instructions";
 
 type Props = {
   amount: number;
@@ -14,7 +14,7 @@ type Props = {
   barcodeContent: string;
   merchantName: string;
   paymentCode: string;
-  instructions: InstructionsType;
+  renderInstructions: () => ComponentChildren;
   title: string;
   t: TFunction;
 };
@@ -27,12 +27,13 @@ export function ActionBarcode(props: Props) {
     onAffirm,
     barcodeContent,
     merchantName,
-    instructions,
+    renderInstructions,
     title,
     t,
   } = props;
 
   const [showSpinner, setShowSpinner] = useState(false);
+  const inActionCard = useActionCard();
 
   const onMadePaymentClicked = useCallback(() => {
     setShowSpinner(true);
@@ -68,12 +69,16 @@ export function ActionBarcode(props: Props) {
 
   return (
     <div className="xendit-action-present-to-customer">
-      <img
-        src={channelLogo}
-        alt="Channel Logo"
-        className="xendit-action-barcode-channel-logo"
-      />
-      <div className="xendit-action-title">{title}</div>
+      {!inActionCard ? (
+        <>
+          <img
+            src={channelLogo}
+            alt={t("image_alt.channel_logo", { channelName: title })}
+            className="xendit-action-barcode-channel-logo"
+          />
+          <div className="xendit-action-title">{title}</div>
+        </>
+      ) : null}
       <div
         data-testid="barcode"
         className="xendit-action-barcode-barcode-container"
@@ -115,7 +120,7 @@ export function ActionBarcode(props: Props) {
           {t("action.payment_confirmation_instructions")}
         </div>
       </div>
-      <Instructions instructions={instructions} />
+      {renderInstructions()}
     </div>
   );
 }