+++ sdk/src/lifecycle/behaviors/action-redirect.ts
@@ -25,7 +25,7 @@ export class ActionRedirectBehavior implements Behavior {
 
   enter() {
     window.addEventListener("pageshow", this.onPageShow);
-    this.bb.dispatchEvent(new XenditWillRedirectEvent());
+    this.bb.dispatchEvent(new XenditWillRedirectEvent(this.url));
     this.bb.telemetry.expectingRedirectAway = true;
     window.location.href = this.url;
   }