+++ sdk/src/lifecycle/behaviors/sdk-fatal-error.ts
@@ -1,3 +1,4 @@
+import { BffErrorContent } from "../../backend-types/common";
 import { XenditFatalErrorEvent } from "../../public-event-types";
 import { BlackboardType } from "../behavior-tree";
 import { Behavior } from "../behavior-tree-runner";
@@ -9,7 +10,16 @@ export class SdkFatalErrorBehavior implements Behavior {
     this.bb.dispatchEvent(
       new XenditFatalErrorEvent(
         this.bb.sdkFatalErrorMessage ?? "Unknown error",
+        this.bb.sdkFatalErrorUserMessage
+          ? errorContentToUserMessage(this.bb.sdkFatalErrorUserMessage)
+          : undefined,
       ),
     );
   }
+}
+
+function errorContentToUserMessage(error: BffErrorContent): string[] {
+  return [error.title, error.message_1, error.message_2].filter(
+    (line) => line !== undefined,
+  ) as string[];
 }